Convert videopress data
