+++ projects/packages/my-jetpack/_inc/data/products/use-activate.ts
@@ -1,14 +1,14 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { REST_API_SITE_PRODUCTS_ENDPOINT } from '../constants';
 import useSimpleMutation from '../use-simple-mutation';
-import useStateProduct from './use-state-product';
+import useProduct from './use-product';
 import type { UseMutateFunction } from '@tanstack/react-query';
 
 const useActivate: ( productId: string ) => {
 	activate: UseMutateFunction;
 	isPending: boolean;
 } = productId => {
-	const { product, refetch } = useStateProduct( productId );
+	const { detail, refetch } = useProduct( productId );
 
 	const { mutate: activate, isPending } = useSimpleMutation(
 		'activateProduct',
@@ -26,7 +26,7 @@ const useActivate: ( productId: string ) => {
 		sprintf(
 			// translators: %$1s: Jetpack Product name
 			__( 'Failed to activate %1$s. Please try again', 'jetpack-my-jetpack' ),
-			product.name
+			detail.name
 		)
 	);
