--- projects/packages/my-jetpack/_inc/data/products/use-all-products.ts
@@ -1,8 +1,0 @@
-const useAllProducts: () => Window[ 'myJetpackInitialState' ][ 'products' ][ 'items' ] = () => {
-	const initialState = window?.myJetpackInitialState;
-	const products = initialState?.products?.items || {};
-
-	return products;
-};
-
-export default useAllProducts;
