+++ projects/packages/my-jetpack/_inc/data/products/use-install-standalone-plugin.ts
@@ -1,14 +1,14 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { REST_API_SITE_PRODUCTS_ENDPOINT } from '../constants';
 import useSimpleMutation from '../use-simple-mutation';
-import useStateProduct from './use-state-product';
+import useProduct from './use-product';
 import type { UseMutateFunction } from '@tanstack/react-query';
 
 const useInstallStandalonePlugin: ( productId: string ) => {
 	install: UseMutateFunction;
 	isPending: boolean;
 } = productId => {
-	const { product, refetch } = useStateProduct( productId );
+	const { detail, refetch } = useProduct( productId );
 
 	const { mutate: install, isPending } = useSimpleMutation(
 		'installPlugin',
@@ -25,7 +25,7 @@ const useInstallStandalonePlugin: ( productId: string ) => {
 		sprintf(
 			// translators: %$1s: Jetpack Product name
 			__( 'Failed to install standalone plugin for %1$s. Please try again', 'jetpack-my-jetpack' ),
-			product.name
+			detail.name
 		)
 	);
